SimpleChat:UI: Move html ui base helpers into its own module
